fix: compare grid canvas fp32 (N14)
The Compose canvas now initialises as torch.float32 instead of
inheriting the first image's dtype.  Previously an fp16 / bf16 input
would force the canvas (and therefore the output) into the smaller
dtype, surprising downstream image nodes that assume fp32.

Tests still pass — the existing compare-grid fixtures use fp32 inputs
so the dtype change is invisible to them.

+++ ComfyUI-LBM-Pro/nodes/lbm_compare_grid.py
@@ -144,7 +144,7 @@ class LBM_Compare_Grid:
         canvas = torch.zeros(
             (rows * H + (rows + 1) * padding,
              cols * W + (cols + 1) * padding, 3),
-            dtype=firsts[0].dtype,
+            dtype=torch.float32,
         )
         for i, im in enumerate(firsts):
             r = i // cols
